Move coordinator setup to integration setup
