+++ homeassistant/components/meater/__init__.py
@@ -1,6 +1,8 @@
 """The Meater Temperature Probe integration."""
+from datetime import timedelta
 import logging
 
+import async_timeout
 from meater import (
     AuthenticationError,
     MeaterApi,
@@ -13,6 +15,7 @@ from homeassistant.const import CONF_PASSWORD, CONF_USERNAME
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
+from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 from .const import DOMAIN
 
@@ -39,10 +42,42 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         _LOGGER.error("Unable to authenticate with the Meater API: %s", err)
         return False
 
-    hass.data.setdefault(DOMAIN, {})
+    async def async_update_data():
+        """Fetch data from API endpoint."""
+        try:
+            # Note: asyncio.TimeoutError and aiohttp.ClientError are already
+            # handled by the data update coordinator.
+            async with async_timeout.timeout(10):
+                devices = await meater_api.get_all_devices()
+        except AuthenticationError as err:
+            raise UpdateFailed("The API call wasn't authenticated") from err
+        except TooManyRequestsError as err:
+            raise UpdateFailed(
+                "Too many requests have been made to the API, rate limiting is in place"
+            ) from err
+        except Exception as err:  # pylint: disable=broad-except
+            raise UpdateFailed(f"Error communicating with API: {err}") from err
 
-    hass.data[DOMAIN][entry.entry_id] = meater_api
-    hass.data[DOMAIN]["entities"] = {}
+        return devices
+
+    coordinator = DataUpdateCoordinator(
+        hass,
+        _LOGGER,
+        # Name of the data. For logging purposes.
+        name="meater_api",
+        update_method=async_update_data,
+        # Polling interval. Will only be polled if there are subscribers.
+        update_interval=timedelta(seconds=30),
+    )
+    await coordinator.async_config_entry_first_refresh()
+
+    hass.data.setdefault(DOMAIN, {})
+    hass.data[DOMAIN].setdefault("entities", {})
+
+    hass.data[DOMAIN][entry.entry_id] = {
+        "api": meater_api,
+        "coordinator": coordinator,
+    }
 
     hass.config_entries.async_setup_platforms(entry, PLATFORMS)
     return True
